Based all the executors on an abstrac class
